rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "llama-index>=0.14.0",
    "llama-index-vector-stores-postgres>=0.6.6",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "youtube-transcript-api>=1.2.2",
]
